Add index & mip-map levels to ShaderUsage.txt
Some of the copies I've seen didn't seem to quite add up, so capture
more information to make sense of it all.

What I'm seeing is resources copied from one mip-map level in one
resource to a different mip-map level in another, and in some cases I'm
even seeing mip-map level 1 copied to level 0, but in most cases mip-map
level 0 is being copied from a different resource... This potentially
complicates the idea of propagating hashes during the copy operations.

The map that tracks the region copies now uses a key that consists of
five values to uniquely identify copies from and to given resources, and
what index and mip-map levels were copied from and to.