Track updates to resources and update/propagate the hash
This is the first attempt to solve the problem of unreliable resource
hashes in Far Cry 4. It adds a new track_texture_updates option to the
d3dx.ini file, which when enabled will try to keep resource hashes in
sync with their contents.

It will recalculate the CRC whenever a resource is updated via a
Map(WRITE*) or an UpdateSubresource() call, and will propagate hashes
from one resource to another on a CopyResource() or CopySubresource()
call. Hashes are only updated/propagated for subresource 0 on the
destination since we are now using the subresource 0 hash for the full
texture.

If UpdateSubresource or CopySubresourceRegion is used to only update a
region of the resource we will not currently update the hash, as doing
so would lead to us using a hash that does not necessarily represent the
full resource. If we were to do anything in this situation, we could
wipe out the data hash on the destination, but it's not clear if that is
the best action - I'd rather wait until we have a proven need to do
something here.

Another edge case which needs further consideration is what to do if a
resource that was created with no initial data is copied into one that
did have. This code will clear out the destination's data hash and
recalculate it's hash based on the desc alone - on the theory that we
will care more about what the resource contains now than what it
contained when it was created... but I'm not positive if that is always
the best answer. We might consider an option to change this behaviour in
the future depending on what we find will be useful in practice.

The CRC of the data is stored separately to the overall texture CRC, and
only the data hash is propagated - the overall texture hash will then be
recalculated using the new data hash as a seed.

The reason to do it this way is because Far Cry 4 does populate it's
resources with an initial texture, and we would calculate it's hash with
usage=default and cpu_access_flags=none. Later it might replace that
texture with a different one, and then later it might replace it with
the original - but when it does that it will have been copied from a
usage=dynamic cpu_access_flags=write resource, so if we propagated the
full hash it would not match the original, even though the resource now
does.

By only propagating the data part of the hash and recalculating the full
hash we will end up with the original hash again (we got lucky that we
had calculated the data hash first otherwise this would not work without
fundamentally changing the hash again).

This patch also splits out all code related to resource hashes into
their own source file and severs the remaining coupling to the Hacker*
objects. It also tweaks some of the includes to break some circular
include dependencies to make this possible.

Needless to say, this change needs to be put through it's paces and
needs it's performance measured to see how much of an impact it will
have in practice. There is the potential to improve performance a little
in the Map() case - currently the code reads the mapped memory to
calculate the GPU, but the documentation indicates that doing so is bad
for performance. We could alternatively divert the mapping returned to
the game to something in system memory, then memcpy it to the real
mapping on the unmap call. It's not clear what the performance of either
approach will be without measuring it though.

TODO: The [TextureOverride] StereoMode and Format settings require the
original hash of a resource - we should track that as well and report it
in the ShaderUsage.txt & frame analysis. In most cases my guess is that
will only be used for resources that are never contaminated with another
hash (mostly render targets).