Only consider 1st subresource in Texture2D hashes
In order to solve the changing texture hashes in Far Cry 4, I intend to
introduce a new feature that changes updates to textures and where
possible updates their hash to match their current contents.

This will involve rehashing resources when they are updated by a Map()
or UpdateSubresource() and propagating hashes when they are copied with
CopyResource() or CopySubresourceRegion().

Analysing the situation has uncovered a problem with this proposed
solution and the way we calculate hashes for 2D textures with multiple
elements in their array. In 3DMigoto 1.2.9 they were changed so that
they might use all elements in the array as part of the hash (but only
if the broken 3DMigoto 1.2.8 hash would overflow the 1st buffer), but
doing so means that if only a single subresource is being updated we
cannot recalculate the hash of the entire texture as we would need to
obtain all subresources to do so.

We can solve this problem by only ever considering the 1st subresource
in the texture hashes - then we only need to update/propagate hashes
when the first subresource is being modified.

This has the potential to change some texture hashes again, but given
that the 3DMigoto 1.2.9 change did not seem to break any fixes, and
given that this change will only affect a subset of those hashes, and
given that multi-element resources are very rarely used we are probably
fine.