Divert D3D11_MAP_WRITE mappings to system memory
The documentation indicates that reading from a writable mapping is bad
for performance as it requires the data to be fetched from the GPU,
which is a high latency operation.

Calculating the new data CRC does require us to read what the game has
written, but it's not clear whether we will suffer this documented
performance bottleneck or not since the documentation lacks low level
hardware details we would need to answer this.

If the buffer is already cached in system memory or the CPU cache we
will be fine since it has just been written and will therefore be hot in
the cache, but if the mapping is in fact an uncacheable IO map that will
involve round trips to the GPU whenever it is read, which will suck for
performance. From my experience working with hardware and the notes in
the documentation, I would guess the later.

An alternative approach is to divert the mapping to system memory that
the CPU can access locally to calculate the CRC, then memcpy it into the
real map when the game calls unmap. This might avoid the latency, but
costs us an extra malloc, memcpy and free.

We really should measure with and without this patch to definitively
answer which is better. Anecdotally I'm not sure - it seemed to stutter
less in FC4 with this patch, but there was some stutter either way, even
when tracking texture updates was disabled.