Resource copying: Add logging when creating any kind of texture
There seem to be a bunch of cases where a resource can't be created with
certain combinations of flags, many of which are non-obvious. Some of
these I've already caught (like creating a resource with a depth/stencil
format with bind flags to allow it to be used in a shader resource), but
there may well be other combinations yet to surprise us.

If we fail to create a resource log the desc structure so we can work
out why, and if debug logging is enabled always log it.

This also changes the debug logging in the various create resource
functions to use the same code, which has the added benefit of decoding
the texture format.