Track original resource hash and switch stat structures to use it
Now that we can update the hash of a resource the data structures we are
using to collect stats on everything can easily get out of sync since
they relied on matching the resource hash, which is no longer
necessarily the case.

Rather than do anything heroic here I'm just tracking the original hash
as well as the current hash, then we can look up these structures by the
original hash instead.

Add the original hash to ShaderUsage.txt and frame analysis filenames so
we have some way to correlate them.

Also tweak how hash contamination is detected to more closely match how
we propagate them (not exact in the case of region copies, but they are
still interesting to track), and disable tracking contamination on
mip-maps again.