Resource copying: Discard stale views
When copying a texture by reference, we will create a view for it and
cache it so we don't have to create a new view every time.

But, if the resource being copied should change (e.g. because the game
settings have been changed), we will continue to use the cached view,
which no longer matches the new resource.

Add a check when using the cached view to make sure it still points to
the resource being copied, and release and recreate it if not.

Reported-by: Oomek