Fix all comparisons to use -1:0 instead of HLSL 1:0 for booleans.
This introduces the cmp helper routines, that will convert the  HLSL boolean of 1:0 into the ASM boolean of -1:0.

This is done as helper routines, to avoid adding inline code into the comparisons, to keep them more clear.

  r1.x = (int)r0.z < 4;
becomes:
  r1.x = cmp((int)r0.z < 4);

Requires the 4 variants of cmp, cmp2, cmp3, cmp4 for the different sizes of possible swizzle. The helper routines are injected into the declaration section.

Tested in FC4 to fix the boolean mismatch in a9c2...