Add version to every Decompiled shader.
This moves the 3Dmigoto declarations to the top of the HLSL shader files, and adds the version and time stamp, so we can keep track of all the different shader files created and what version created them.

This is prompted by the addition of the cmp routine section, seems past time to version stamp the output files.