Return error for HackerDXGIAdapter::CheckInterfaceSupport.
This matches the goal of returning errors for any devices we don't support, anything except DX11.

This returns the DXGI_ERROR_UNSUPPORTED for anything except ID3D11Device, which will be pretty much always.  This call is only used for old Vista support or other bad tests, it's not supposed to be used anymore.

This fixes a crash in FarCry4 on Win7, when no evil-update is installed.