Far Cry 4: Remove broken water shader - river in prologue
This looks like an early attempt of Mike's at fixing the water (which
ended up being fixed elsewhere). It currently damaged the ripples and
foam and the original shader did not need to be adjusted, so remove it
from the fix.

This water does have one remaining glitch when MSAA or TXAA is enabled -
an outline of the current weapon appears to the right (noticeable with
the knife, or while running into the sunlight with the gun), however
this appears even in 2D with 3DMigoto removed, so I'm dismissing that as
a game bug.