Make command list parsing completely case insensitive
Command list parsing was already partially case insensitive, but only
for all lvalues other than oD, and all only for resource rvalues.

The ini file is supposed to be case insensitive, so for consistency
convert both lvalue and rvalue operands to lower case before calling any
of the parsing functions, and remove the few special cases they already
had.

This fixes not being able to assign a depth target.