Add info line drawing below shader counts in OSD.
A new line of text, that will be pulled from the HLSL files, to assist ShaderHackers in whatever way they find it helpful.  Plan is to pull first line out of the file, and show it here.

This presently is a stub, showing "test test test" for the line, but the infrastructure for drawing and finding the info text is in place and working here.

This adds another element to the OriginalShaderInfo of "infoText" which will contain that first line from the shader file, and created the same time the overriden shader is loaded.