Tweak the header output for blank comment line.
Plan is to use the first line of HLSL code as on screen comments during hunting.  So, this moves the header down one line so it doesn't always show up.

Since this is new, there should be no conflict with old shaders.

Also tweak comment text for slightly better clearing the string, in the no-comment case.