Add keybinding to toggle forcing full screen
This is essentially a hack to make it easier to work with Unity games.
Unity has had a long standing bug in their DirectX 11 exclusive full
screen implementation, where they will crash, hang or otherwise fail to
recover after alt+tabbing out.

This happens regardless of whether the game was shipped using this mode,
or whether we have forced it with 3DMigoto. It can be reproduced easily
by building even the most trivial test scene in Unity set to use DX11
and exclusive mode full screen.

Of course, we need to use exclusive mode full screen to get 3D to engage
so unless Unity fixes their bug we have to live with it.

As it turns out, their alt+tab handling may be broken, but their
alt+enter handling works just fine to switch the game to windowed mode,
and from there alt+tab is safe to use.

The problem is, if we have forced a game to use exclusive mode full
screen with 3DMigoto to get 3D to engage, we cannot use alt+enter to
switch to windowed since 3DMigoto will just force it straight back to
exclusive mode full screen.

We can work around this by changing the full_screen setting in d3dx.ini,
reloading the config live then switching to windowed mode. Of course,
this requires us to be able to edit the config, which we can't do in
full screen mode (at least, not without a second computer).

This patch adds a key to toggle forcing exclusive mode full screen on
and off. Pressing the key won't have any immediate effect, but it allows
us to then follow it up with alt+enter to switch the game to windowed
mode. This should make it a lot easier to work with Unity games.