Show first line of text of shader file in overlay.
Should work for ASM and HLSL both.  If a file is moved to ShaderFixes either manually or by CopyOnMark, it will get added to the OriginalShaderInfo for mReloadedShaders.

By default the first line of the file will be blank, so it's up to anything the ShaderHacker wants to put there.

Only tested in FC4, but both load at start, and F10 reload worked.