Far Cry 4: Re-enable map vignette adjustment
Now that we have texture hash tracking in 3DMigoto, we should be able to
match the vignette hashes reliably, so re-enable the adjustment.

This adds both texture detection and a full screen check, but it only
checks the x coordinates (of all four corners), since one of the two
vignette textures is just a band on the top/bottom of the screen and not
a complete full screen effect.