Far Cry 4: Fix repair tool flame
This uses 3DMigoto's new ability to run command lists from
TextureOverride sections to act as a limited form of scene detection. It
sets a flag when it detects the repair tool's texture used in a frame,
and will clear the flag at the start of the next frame. Since the
misaligned flame is drawn after the opaque repair tool the flag will be
set and it can apply the depth offset to realign the flame.

It will still break unrelated flames, but only while the repair tool is
being used, which should be rare enough that it is acceptable.