Key overrides: Add "condition" option to work with scene detection
This adds a condition option to the key overrides, which can be set to
any of the IniParams and will prevent that override from being activated
if the relevant IniParam is zero.

If this is combined with scene detection (even just the limited form we
have today), it could be used to make a convergence override only
activate while a weapon is being drawn on the screen, or to prevent it
from activating when something else in the scene indicates that the key
is being used for a different action.

This is fairly basic for now - it only checks a single IniParam for
zero/non-zero and the check is only performed on activation of the
override (as opposed to continuously checking once per frame while the
override is active).