FC4: Disable aiming convergence override while descending ropes
Far Cry 4 uses the same key binding to descend ropes as it does to aim
down the sights. Since we use a convergence override on this key, this
results in the world getting further away while climbing down ropes,
which is a rather odd sensation.

We use the new 3DMigoto feature to make convergence overrides
conditional to disable the adjustment whenever the vertex shader for the
climbing rope is active in the scene.

It is still possible to hold the RMB at other times when not using a
weapon with aim down sights (e.g. with the repair tool out, while in a
no fire zone, cutscene, while reloading, etc), but in those cases the
button does nothing so we don't really care. We could potentially match
on all possible weapons instead, but given they can all be skinned in a
million ways and the enemies also carry weapons that doesn't seem
sensible.