Far Cry 4: Add note to icon filter texture
One of the icons is used in two circumstances, one of which should be
adjusted and the other should not be (at least, not without adjusting
several other icons and text - which I don't want to risk now since I
want to get the update out).

The one that is being adjusted that shouldn't be is just the background
to the circle with the XP awards for various actions. For the most part,
the background circle is dark and would usually be further back than the
rest of the circle, so is probably acceptable to leave most of the time.
The circle used while grappling seems more noticeable when it is at the
wrong depth, so leave it set for that, and add a note in the texture to
remind me if I ever revisit this later.