Far Cry 4: Separate XP icon background from grapple target icon
The XP award icon is always displayed on the left outside of the
region of the screen used for floating icons, while the grapple
targeting circle should always be close to the center of the screen.
Paint the icon in the icon filter texture with a special colour and
use that to check which region of the screen it is in to determine which
adjustment to apply.