Allow StereoParams and IniParams to be assigned to different slots
This adds options to the d3dx.ini to allow StereoParams and IniParams to
be assigned to different shader resource slots (or not at all), which
can allow 3DMigoto to be used in games that already use the default
slots, such as DCS World.

Currently it is still only a global setting - we could also consider
allowing it to be assigned on a per-shader basis via the arbitrary
resource copying code.

It's also now possible to not assign it at all, which might be useful
for modders who are not using 3DMigoto for stereo, or by using nvidia's
stereo parameters found in vs-cb12 instead (not recommended).

Requested-by: lefuneste