Batman: Fix view-space halo on officer's face
This is pretty much just a halo type issue, but the output position has
been copied in a different coordinate system (not view-space, probably
world-space), so we need to apply the correction to a copy of the output
position and use the inverse matrix to correct the copy the game made.

Add a matrix.hlsl with an inverse function so we can keep the shaders
cleaner.