Frame Analysis: Fix filenames of constant buffers and textures from compute shaders
Dumping constant buffers and textures from compute shaders with frame
analysis did not pass the fact that they were from compute shaders down
to the filename routine, so it would generate them with the hashes of
the currently set vertex shaders, pixel shaders, etc. instead of the
current compute shader

Pass the compute flag down to fix the filename.