Frame Analysis: Add logging infrastructure
This adds a new frame analysis option "log" which will be used to log
all draw calls and state changes from the frame to a file. This will be
useful to see some information that is not revealed in the existing
dumps, either because it cannot be dumped, or is something that we do
not currently dump (e.g.  a resource copy operation, map, etc).

It will also mean that we have a way to quickly get a list of shaders
used in a frame without having to dump out all the render targets.

If deferred contexts are used, they will each get their own log file,
though at the moment their logging will be tied to the start and end of
the frame from the POV of the immediate context - so they could miss
calls if the game started writing to a deferred context to be used in a
future frame.

Log files are opened on demand and closed once frame analysis is
disabled, or the context released.

This first commit just includes the infrastructure for this support -
the actual logging will be added in subsequent commits to make the
review process easier.