Frame Analysis: Log Get and Set ConstantBuffer calls
This adds logging to all calls that get or set constant buffers. The
calls are logged, as are each of the (non-NULL) constant buffer slots
being accessed.

Set calls are logged at the start of the function, Get calls are logged
at the end.

ID3D11DeviceContext1 variants are also logged - we don't currently allow
these to be used, but this should help debugging in the future if we do
enable it.