Frame Analysis: Log calls to set/get render targets & UAVs
This will log the call as well as all RTV/UAV/DSV slots that are
accessed.

Set calls are logged at the start of the funcion, Get calls at the end.
Get calls check if the depth/stencil view was requested to avoid
dereferencing a NULL pointer since it is the only one that will not be
checked by the Frame Analysis infrastructure.