Frame Analysis: Log internal messages to frame log
Since we now have a log file for the frame being analysed, it is the
logical place to log any info or errors encountered during frame
analysis. We still might want to see these in the main log file as well,
so use a macro to direct these messages to both.