Frame Analysis: Use mOrigContext where applicable
The frame analysis code calls back into the HackerContext in a few
places, which will now cause those calls to be in turn logged into the
frame analysis log. These calls don't need to go through the
HackerContext, so call them on the original context instead.

This will also make it easier to decouple the frame analysis from the
HackerContext later on.