Pass nvapi mOrigDevice instead of the HackerDevice
The frame analysis log was showing a lot of noise in the form of calls
to Begin() and End(), which turned out to be due to nvapi calling them
on the immediate context of the device we used to create the stereo
handle.

Currently we have no reason to intercept calls from nvapi into the
context, so change it to use the original device to keep the log clean.
If we later determine that we need to use the hacker device we can
change this back easily enough.