Batman: Lighting grid WIP
This is a work in progress to try to fix the grid problem with the
lights. I believe that the t10, t13 and/or t17 structured buffers hold
relevant information that the shader uses to decide which lights to draw
in each part of the grid. I have determined with frame analysis that
these structured buffers are populated from the CPU (as opposed to
another shader on the GPU) so we have to work with the information we
have.

My current approach is to try to enable all lights in each part of the
grid. I'm faking the value read from t10[4096 + threadID] to enable 32
lights in each grid (any more introduces flickering), and disabling one
of the conditions elsewhere in the shader to convince it to draw lights
that were not in the grid in 2D but now are in 3D.

This is a significant improvement, but there is still some clipping
problems, so it is not the final solution.

I have also been comparing this shader to the source code in UE4 to try
to reconcile the two, but the usf code in UE4 is hard to follow and
there are a number of differences here (such as the size and layout of
the structured buffers) that make think I may not be looking at the same
shader.

Examining the raw structured buffers dumped with frame analysis does not
match my expectations from the code, so I may be missing something or be
looking at the wrong offsets.