Batman: Add shader names identified with extract_unreal_shaders.py
The script currently only extracts the global shader names, of which we
have only fixed a few. Material shader names are not currently
retrieved as they are stored elsewhere.

This confirms that the shadows in this game are not standard UE4
shadows, as the shader I am looking at is in fact
"FTileShadingComputeShaderTSCS_Wetness00NVOPS_KeplerOptimised", which is
not in the UE4 source anywhere.