Fix off by one error in assembly literal fixup code creating double negatives
In batman we saw some assembly code being produced like this:

          mov r19.yzw, l(0,--0.444036931,00.0614555776,--0.229850411)
          mov r23.xyz, l(-0.255772650,--2.148475353E-008,--0.0497171432,0)
          mov r24.xyz, l(-0.888073862,--7.763788545E-008,--0.459700823,0)
          mov r25.xyz, l(-0.888073862,--7.763788545E-008,--0.459700823,0)

Which clearly contains some issues - the first character on the 2nd
through last literals are duplicated, especially noticeable as it
creates double negatives. Comparing the original shader binary with the
reassembled binary shows that the originals were only single negatives
and the reassembled binary replaces these with 0s.

This was just an off by one error in some copy & pasted code that was
not adjusted for the length of the string being matched.

Closes #37