Resource Copying: Copy format, offset, stride and size from structured buffers
If copying from a structured buffer we need to copy the stride, format,
offset and buffer size so that we can create a new view of the resource
(and optionally do the region copy) later.

Move the responsibility for gathering these pieces of information from
the resource or view descriptions into a central function and add code
to get the information from the view. Some of the information may still
be retrieved earlier if it was retrieved by the call that retrieved the
resource.