Frame Analysis Log: Log draw/dispatch calls after the Before call
I want to add some logging to the command lists which may be called from
the BeforeDraw/BeforeDispatch calls, so to make sure that the log makes
sense, move the logging of the draw/dispatch calls to after that point.
That way commands in the pre list will show before the draw/dispatch
call and commands in the post list will show afterwards.

Also moves the FrameAnalysisAfterDraw to after the post command list has
run so that the draw call numbers will make sense in that command list.
Slight functional change as skipping a shader will now skip the frame
analysis, but if we are skipping a shader we probably don't really care
anyway (and it will still be in the log.txt).