Command List: Add frame analysis logging to each run() function
This will log each command list command in the frame analysis log when
it is run. These are clearly distinguished from log events from the game
by the 3DMigoto prefix. Detail is fairly minimal - since the command
list calls the original context directly for any state changes it makes
a lot of those details won't be logged, but at least we will see the
commands that affected the state.