Command List: Log failed attempts to create a view description
We can't always create a view with a NULL description as that only works
under certain situations, and outside of those situations we will need
to create a view ourselves. If that fails it will be handy to know why,
so log the view description we tried to use and the resource description
if it fails, or whenever debug logging is enabled.