Add ability to hook deferred contexts instead of wrapping them
This solves issues with MGSV:GZ, which exited if we handed it a deferred
context with a vtable that it wasn't expecting (in particular, if any of
the DeviceChild or DeviceContext entries did not match DirectX -
IUnknown entries could be overridden, and immediate contexts and the
device itself did not matter).

This game performs all it's rendering in deferred contexts (which will
unfortunately limit frame analysis' usefulness), so not wrapping the
deferred contexts was not an option.

This adds a new HookedContext.cpp with the infrastructure to hook
contexts instead of wrapping them. It includes two sets of trampoline
functions - the first set is used to intercept calls from the game and
will call into a HackerContext corresponding to a device context and the
second set is used to call the function on the original context.

The first set will be called via hooks and will look up the
HackerContext from an unordered_map. The second set is packaged up into
a new COM object implementing the ID3D11DeviceContext interface and
handed back to the HackerContext to use in place of the original
context.

This design allows us to integrate hooking with no changes to the
HackerContext beyond a simple set up function. This means no duplicated
code and no copy and pasted bugs.

A new setting is added to the ini to enable the hooks:

[System]
hook = deferred_contexts

Currently only "deferred_contexts" and "all" is supported, but it is
intended that we can add to this list as hook more. I considered a
simple on/off switch, but given that hooking adds several extra layers
of indirection and given that it is only necessary in some special cases
and given that we don't necessarily know how other games will react to
it I felt it better to give us more fine grain control if needed. The
"all" keyword is intended to be used as a simple way to enable all
hooks (once we have more anyway).