Fix logging proxy_d3d11 & remove needless checks on LogInfo
The "if (!G->CHAIN_DLL_PATH)" test would never pass to log the proxy
dll, so change it to match the same form as other similar settings.

Remove several if (LogInfo) tests - this function is generally only run
once and is not performance critical, so it's better to be more readable
rather than pre-checking that.