HookedContext: Add locking around context_map
It's not super clear that I need locking on the context_map - deferred
contexts have to be created from the immediate context, so that would
suggest that it should be single threaded. However we do remove entries
from the map when refcounts reach 0 and that could cause problems if
another thread was looking up the map, so add locking to be safe.

We don't add any locking on the trampoline to call back to the game -
since they are allocated per HackerContext they should never be accessed
from multiple threads simultaneously.