Add HookedDevice & hook immediate context
This works exactly the same as the HookedContext, but for devices.

MGS:TPP seems to get a bit further with this, but the driver hits a null
pointer dereference (NULL + 0x28) which I have yet to resolve.

Since this feature has to be explicitly enabled this should not impact
anyone yet.