Hooks: Fix crashes in CreateSwapChain being unable to find the HackerDevice
CreateSwapChain called typeid on the device, which now may be the
original device which would crash due to not containing any RIID
information.

Export a function from the hooking infrastructure that can be used to
look up a HackerDevice from a hooked device, and use that if the RIID
information is missing.