Prevent ever using typeid on hooked device/context to fix crash
It seems that typeid is even less safe than I thought - if it is called
on something without RIID it may raise an exception, or it might just
crash the program as it looks at memory that isn't mapped in, and the
type_name wrapper only helps in the former case.

Check if the object being passed in is a hooked device/context before
calling typeid in the type_name wrapper, as those should be the only
objects that we might call it on that do not have RIID info.

Also change CreateSwapChain to also check for hooked devices/contexts
first and only use typeid if it was not found.