Hooking: GetOrigContext/Device always return the real original object
Trying to sort out what uses the trampoline objects and what uses the
real original objects.

It seems that most things that call GetOrigBlah will want the real
originals, though some of them might want the trampolines to avoid
calling back into 3DMigoto, but for now switch everything to use the
real originals and we can change any code to use the trampolines
explicitly as we think is necessary.

Still seeing crashes when using hook=all in Witches 3, but seemingly not
when only hooking either device or context. The crashes have a very
consistent stack and instruction pointer, but it very consistently
doesn't make sense. Forcing CPU affinity seems to solve the crash (as it
does in MGSV:GZ) suggesting that there is a threading issue with either
the hooks, or the code I run from them. Notably, Witcher 3 continually
creates and destroys deferred contexts, which would fit with the
threading theory.

MGSV:TPP still crashes at launch in the driver, but it's not yet clear
if that is related to the hooking or just the next independent problem
to solve.

FC4 seems perfectly stable using hook=all.