Resource copying: Fix regression copying index buffers
When refactoring the code to fill in missing information for buffers, I
missed the case of copying a non-structured buffer that has a format but
no view, such as an index buffer or another custom resource. This would
cause a divide by zero error when creating a view for these crashing
3DMigoto.

Fix the issue by filling in the stride based on the DXGI format if there
is no higher priority means of determining it.