Resource Copying: Take per-draw offset & length into account for vertex & index buffers
The code to copy vertex and index buffers will take the byte offset
assigned when the buffers are bound into account, but will not consider
the offset and length passed into the current draw call.

For some games / effects like the UI in Far Cry 4 this is fine, as they
only use the offset when binding a buffer to the pipeline, but for other
games / effects like the UI in Witcher 3 this does not work as all calls
have an offset.

Refactor how info from the draw calls is passed around a little and pass
the first vertex/index/instance information down to the command list
which can take it into account when copying from a vertex or index
buffer.

If the draw call does not contain this information it will be obtained
from the buffers themselves like before assuming there is no per-draw
offset or length.