Frame Analysis: Vertex & Index buffer dumping improvements
Pass the draw info into the frame analysis code and use it in
dump_vb_txt and dump_ib_txt to only dump the parts of the buffers
actually used in the draw call when known.

Change dump_vb_txt to use a dedicated function that dumps the buffer as
hex and floats and takes the stride into account. We still don't have a
good way to decode the vertex buffers correctly, but this should give
more useful information than using the generic constant buffer dump
routine.