HLSL Decompiler: Add ibfe instruction
This instruction is used by recent versions of Dolphin. It's similar to
the PPC rwlninm instruction used to extract a field from a register,
except it's sign extending and can't be used for arbitrary rotations.

Not sure if there's a clean way to write the instruction in HLSL, so for
now just write it using a ternary operator and bring back the swizCount
helper. Seems to generate the correct assembly (except with an extra
round_z).

Reported-by: masterotaku