HookedContext: Work around crash in XXSetShaderResources when NumViews=0
This is very strange - when we hook VS/PSSetShaderResources we get
crashes in MGSV:TPP. Logging shows that these occur when we call the
original SetShaderResources call via orig_vtable, but only when
NumViews=0.

For now, work around the issue by bailing when NumViews=0. Alternatively
we could just not hook these calls, but I don't want to do that without
understanding the root cause as much as possible in case it is a wider
problem with hooking (such as stack corruption).