HookedContext: Add except_set_shader_resources option
I don't want to remove the logging about the NumViews=0 problem until we
understand it, but... it makes for a very large log file.

Add an option to explicitly not hook the calls to avoid the log growing
too much, while still acknowledging that there is a problem.

Had to refactor the enable hooks definition a little to enable this, as
the Hooked*.cpp files cannot include Globals.h directly thanks to their
use of the C API in D3D11.h.